Implement the Web memory API for service/shared workers
This enables performance.measureUserAgentSpecificMemory in service and
shared workers guarded by cross-origin isolation.

A service worker is not allowed to spawn a dedicated worker and Chrome
does not support dedicated workers nested in shared workers. This means
that the web memory API needs to measure only the current execution
context. This allows a simple local implementation without doing a
round-trip to PerformanceManager in the browser process.

Cross-origin isolation for service and shared workers is currently
being implemented (crbug.com/1131404 and crbug.com/996511). The tests
are disabled until that work completes.

Bug:326370
Change-Id: I1a49113c058b44c3dd35e8b1fb97fe5be39fbd8d