[LayoutNG] Allow spanners inside a transform
Previously, CanContainSpannerInParentFragmentationContext() would not
allow spanners that are the descendant of a transform. However, the
containing block chain now goes directly from the column spanner
to the multicol container for an NG multicol. Thus, it is safe to
allow a spanner to be a true column spanner if it is a descendant of
a transform (in the case that the multicol is an NG object).

The web test that was added to test the old behavior has been removed
and a new test has been added to test the new behavior.

This CL causes a new test to fail. This only fails in combination with
CL:2870868. See the comment in TestExpectations for more details:
virtual/layout_ng_block_frag/external/wpt/css/css-multicol/filter-with-abspos.html

Bug: 1196727
Change-Id: I908d4fdfc6ce090187dedbc2d9488b1c09bfa989